FIX: - one view for popup - move getFilmComments from view to movie presenter
